feat: Update Long to Integer for all Ids
Updated all entity, DTO, repository, service, and controller files to use Integer instead of Long for all Ids.
This change was made to ensure consistency across the application.

+++ src/main/java/com/management/financial/controller/CostCenterController.java
@@ -26,17 +26,17 @@ public class CostCenterController {
     }
 
     @GetMapping("/{id}")
-    public CostCenterDTO getCostCenterById(@PathVariable Long id) {
+    public CostCenterDTO getCostCenterById(@PathVariable Integer id) {
         return costCenterService.getCostCenterById(id);
     }
 
     @PutMapping("/{id}")
-    public CostCenterDTO updateCostCenter(@PathVariable Long id, @RequestBody CostCenterDTO costCenterDTO) {
+    public CostCenterDTO updateCostCenter(@PathVariable Integer id, @RequestBody CostCenterDTO costCenterDTO) {
         return costCenterService.updateCostCenter(id, costCenterDTO);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCostCenter(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteCostCenter(@PathVariable Integer id) {
         costCenterService.deleteCostCenter(id);
         return ResponseEntity.noContent().build();
     }
